Separate header text into its own component
